[75] Add React Router and refactor app component
