+++ package.json
@@ -11,6 +11,7 @@
     "node-sass": "4.13.1",
     "react": "16.13.0",
     "react-dom": "16.13.0",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0"
   },
   "scripts": {
