+++ src/main/js/SurveyPage.scss
